Fix issue #149
When running on an extremely show machine, it's possible that multiple
timer callbacks can overlap. When that happens we're supposed to just
run one, but this code incorrectly deals with that case:
- one of the callbacks will run correctly
- one of the callbacks will throw SynchronizationLockException

The problem here is that the `Monitor.TryEnter` is inside the
try/finally that releases the lock. Oops.